refactor(mlx-lm): unify family layout under crate::<family>::text/
Restructure llama + qwen3 into one per-family layout so qwen3.5 (and
future families) drop in identically. Pure relocation + dispatch
convention — model graphs and internals unchanged, decode path
untouched, bench-neutral.

- models/<family>.rs -> <family>/text/model.rs; <family>/adapter.rs ->
  <family>/text/adapter.rs; ModelArgs split into <family>/text/config.rs.
- Family: name() + as_llama()/as_qwen3(). Adapters take (cfg, args, dir)
  and stop re-reading config; load_context(cfg, dir) dispatches via
  as_<family>(). model_context::dispatch_load passes cfg through.
- Drop the models/ module (code ported, not removed).
