+++ mlx-lm/src/llama/text/mod.rs
@@ -1,0 +1,5 @@
+//! Llama text path: config, model graph, and the runtime adapter.
+
+pub mod adapter;
+pub mod config;
+pub mod model;
