+++ mlx-lm/src/llama/text/model.rs
@@ -20,36 +20,14 @@ use crate::{
     cache::KeyValueCache,
     config::{Family, ModelConfig},
     error::Error,
-    family::EosSpec,
+    llama::text::config::ModelArgs,
     loader::apply_post_load_memory_policy,
     nn::{ensure_cache_populated, AttentionInput, ModelInput, SwigluMlp},
     utils::{
-        create_attention_mask,
-        rope::{initialize_rope, FloatOrString, RopeVariant},
-        scaled_dot_product_attention, AttentionMask,
+        rope::{initialize_rope, RopeVariant},
+        scaled_dot_product_attention,
     },
 };
-
-/// Qwen3 config body. `model_type` is the serde tag on
-/// [`crate::config::Family`]; quantization lives on the outer
-/// [`crate::config::ModelConfig`].
-#[derive(Debug, Clone, Deserialize)]
-pub struct ModelArgs {
-    pub hidden_size: i32,
-    pub num_hidden_layers: i32,
-    pub intermediate_size: i32,
-    pub num_attention_heads: i32,
-    pub rms_norm_eps: f32,
-    pub vocab_size: i32,
-    pub num_key_value_heads: i32,
-    pub max_position_embeddings: i32,
-    pub rope_theta: f32,
-    pub head_dim: i32,
-    pub tie_word_embeddings: bool,
-    pub rope_scaling: Option<HashMap<String, FloatOrString>>,
-    #[serde(default)]
-    pub eos_token_id: Option<EosSpec>,
-}
 
 #[derive(Debug, Clone, ModuleParameters, Quantizable)]
 pub struct Attention {
@@ -70,10 +48,6 @@ pub struct Attention {
     #[param]
     pub o_proj: MaybeQuantized<nn::Linear>,
     #[param]
-    pub q_norm: nn::RmsNorm,
-    #[param]
-    pub k_norm: nn::RmsNorm,
-    #[param]
     pub rope: RopeVariant,
 }
 
@@ -87,23 +61,16 @@ impl Attention {
         let scale = (head_dim as f32).sqrt().recip();
 
         let q_proj = nn::LinearBuilder::new(dim, n_heads * head_dim)
-            .bias(false)
+            .bias(args.attention_bias)
             .build()?;
         let k_proj = nn::LinearBuilder::new(dim, n_kv_heads * head_dim)
-            .bias(false)
+            .bias(args.attention_bias)
             .build()?;
         let v_proj = nn::LinearBuilder::new(dim, n_kv_heads * head_dim)
-            .bias(false)
+            .bias(args.attention_bias)
             .build()?;
         let o_proj = nn::LinearBuilder::new(n_heads * head_dim, dim)
-            .bias(false)
-            .build()?;
-
-        let q_norm = nn::RmsNormBuilder::new(head_dim)
-            .eps(args.rms_norm_eps)
-            .build()?;
-        let k_norm = nn::RmsNormBuilder::new(head_dim)
-            .eps(args.rms_norm_eps)
+            .bias(args.attention_bias)
             .build()?;
 
         let rope = initialize_rope(
@@ -122,8 +89,6 @@ impl Attention {
             k_proj: MaybeQuantized::Original(k_proj),
             v_proj: MaybeQuantized::Original(v_proj),
             o_proj: MaybeQuantized::Original(o_proj),
-            q_norm,
-            k_norm,
             rope,
         })
     }
@@ -131,7 +96,7 @@ impl Attention {
 
 impl<C> Module<AttentionInput<'_, C>> for Attention
 where
-    C: KeyValueCache + Default,
+    C: KeyValueCache,
 {
     type Output = Array;
 
@@ -149,16 +114,12 @@ where
         let keys = self.k_proj.forward(x)?;
         let values = self.v_proj.forward(x)?;
 
-        let mut queries = self.q_norm.forward(
-            &queries
-                .reshape(&[B, L, self.n_heads, -1])?
-                .transpose_axes(&[0, 2, 1, 3])?,
-        )?;
-        let mut keys = self.k_norm.forward(
-            &keys
-                .reshape(&[B, L, self.n_kv_heads, -1])?
-                .transpose_axes(&[0, 2, 1, 3])?,
-        )?;
+        let mut queries = queries
+            .reshape(&[B, L, self.n_heads, -1])?
+            .transpose_axes(&[0, 2, 1, 3])?;
+        let mut keys = keys
+            .reshape(&[B, L, self.n_kv_heads, -1])?
+            .transpose_axes(&[0, 2, 1, 3])?;
         let values = values
             .reshape(&[B, L, self.n_kv_heads, -1])?
             .transpose_axes(&[0, 2, 1, 3])?;
@@ -184,8 +145,6 @@ where
         self.k_proj.training_mode(mode);
         self.v_proj.training_mode(mode);
         self.o_proj.training_mode(mode);
-        self.q_norm.training_mode(mode);
-        self.k_norm.training_mode(mode);
         <RopeVariant as Module<nn::RopeInput>>::training_mode(&mut self.rope, mode);
     }
 }
@@ -216,7 +175,7 @@ impl TransformerBlock {
         let hidden_size = args.hidden_size;
 
         let self_attn = Attention::new(args)?;
-        let mlp = SwigluMlp::new(args.hidden_size, args.intermediate_size, false)?;
+        let mlp = SwigluMlp::new(args.hidden_size, args.intermediate_size, args.mlp_bias)?;
         let input_layernorm = nn::RmsNormBuilder::new(args.hidden_size)
             .eps(args.rms_norm_eps)
             .build()?;
@@ -237,7 +196,7 @@ impl TransformerBlock {
 
 impl<C> Module<AttentionInput<'_, C>> for TransformerBlock
 where
-    C: KeyValueCache + Default,
+    C: KeyValueCache,
 {
     type Output = Array;
 
@@ -269,7 +228,7 @@ where
 }
 
 #[derive(Debug, Clone, ModuleParameters, Quantizable)]
-pub struct Qwen3Model {
+pub struct LlamaModel {
     pub vocab_size: i32,
     pub num_hidden_layers: i32,
 
@@ -285,7 +244,7 @@ pub struct Qwen3Model {
     pub norm: nn::RmsNorm,
 }
 
-impl Qwen3Model {
+impl LlamaModel {
     pub fn new(args: &ModelArgs) -> Result<Self, Exception> {
         assert!(args.vocab_size.is_positive());
 
@@ -310,7 +269,7 @@ impl Qwen3Model {
     }
 }
 
-impl<C> Module<ModelInput<'_, C>> for Qwen3Model
+impl<C> Module<ModelInput<'_, C>> for LlamaModel
 where
     C: KeyValueCache + Default,
 {
@@ -329,13 +288,15 @@ where
 
         let mask = match mask {
             Some(mask) => Some(mask.clone()),
-            None => match create_attention_mask(&h, cache, Some(true))? {
-                Some(AttentionMask::Array(a)) => Some(a),
-                Some(AttentionMask::Causal) => {
-                    return Err(Exception::custom("Only `Array` mask is supported"));
+            None => {
+                if h.shape()[1] > 1 {
+                    let m =
+                        nn::MultiHeadAttention::create_additive_causal_mask::<f32>(h.shape()[1])?;
+                    Some(m.as_dtype(h.dtype())?)
+                } else {
+                    None
                 }
-                None => None,
-            },
+            }
         };
 
         ensure_cache_populated(cache, self.layers.len());
@@ -367,7 +328,7 @@ pub struct Model {
 
     #[quantizable]
     #[param]
-    pub model: Qwen3Model,
+    pub model: LlamaModel,
 
     #[quantizable]
     #[param]
@@ -376,7 +337,7 @@ pub struct Model {
 
 impl Model {
     pub fn new(args: ModelArgs) -> Result<Self, Exception> {
-        let model = Qwen3Model::new(&args)?;
+        let model = LlamaModel::new(&args)?;
         let lm_head = if !args.tie_word_embeddings {
             Some(MaybeQuantized::Original(
                 nn::LinearBuilder::new(args.hidden_size, args.vocab_size)
@@ -416,14 +377,14 @@ where
     }
 
     fn training_mode(&mut self, mode: bool) {
-        <Qwen3Model as Module<ModelInput<'_, C>>>::training_mode(&mut self.model, mode);
+        <LlamaModel as Module<ModelInput<'_, C>>>::training_mode(&mut self.model, mode);
         if let Some(lm_head) = &mut self.lm_head {
             lm_head.training_mode(mode);
         }
     }
 }
 
-pub fn load_qwen3_tokenizer(model_dir: impl AsRef<Path>) -> Result<Tokenizer, Error> {
+pub fn load_llama_tokenizer(model_dir: impl AsRef<Path>) -> Result<Tokenizer, Error> {
     let file = model_dir.as_ref().join("tokenizer.json");
     Tokenizer::from_file(file).map_err(Into::into)
 }
@@ -434,12 +395,12 @@ pub struct WeightMap {
     pub weight_map: HashMap<String, String>,
 }
 
-pub fn load_qwen3_model(model_dir: impl AsRef<Path>) -> Result<Model, Error> {
+pub fn load_llama_model(model_dir: impl AsRef<Path>) -> Result<Model, Error> {
     let model_dir = model_dir.as_ref();
     let cfg = ModelConfig::from_dir(model_dir)?;
     let quant = cfg.quantization().cloned();
-    let Family::Qwen3(args) = cfg.family else {
-        return Err(Error::config("config.json model_type is not qwen3"));
+    let Family::Llama(args) = cfg.family else {
+        return Err(Error::config("config.json model_type is not llama"));
     };
     let mut model = Model::new(args)?;
     if let Some(q) = quant {
@@ -493,19 +454,134 @@ pub fn load_qwen3_model(model_dir: impl AsRef<Path>) -> Result<Model, Error> {
 
 #[cfg(test)]
 mod tests {
-    use mlx_rs::{module::Module, transforms::eval, Array};
+    use std::{env::home_dir, fs};
 
-    use crate::{cache::KVCache, models::qwen3::load_qwen3_tokenizer};
+    use lazy_static::lazy_static;
+    use mlx_rs::{transforms::eval, Array};
 
-    const CACHED_TEST_MODEL_DIR: &str = "../cache/Qwen3-4B-bf16";
+    use crate::{
+        cache::KVCache,
+        config::{Family, ModelConfig},
+        llama::text::model::{load_llama_model, load_llama_tokenizer},
+    };
+
+    /// Resolve the HuggingFace cache directory to the actual snapshot path.
+    /// The structure is:
+    ///   models--<org>--<name>/
+    ///     refs/
+    ///       main  (contains the commit hash)
+    ///     snapshots/
+    ///       <commit_hash>/  (actual model files)
+    fn resolve_hf_cache_dir(model_cache_dir: &str) -> String {
+        let refs_main = std::path::Path::new(model_cache_dir)
+            .join("refs")
+            .join("main");
+        let commit_hash = fs::read_to_string(&refs_main)
+            .unwrap_or_default()
+            .trim()
+            .to_string();
+        std::path::Path::new(model_cache_dir)
+            .join("snapshots")
+            .join(commit_hash)
+            .to_string_lossy()
+            .into_owned()
+    }
+
+    lazy_static! {
+        static ref CACHED_TEST_MODEL_DIR: String = {
+            let cache_dir = home_dir()
+                .map(|p| {
+                    p.join(".cache")
+                        .join("huggingface")
+                        .join("hub")
+                        .join("models--meta-llama--Llama-3.2-1B-Instruct")
+                        .to_string_lossy()
+                        .into_owned()
+                })
+                .unwrap_or_default();
+
+            resolve_hf_cache_dir(&cache_dir)
+        };
+        static ref CACHED_QUANT_TEST_MODEL_DIR: String = {
+            let cache_dir = home_dir()
+                .map(|p| {
+                    p.join(".cache")
+                        .join("huggingface")
+                        .join("hub")
+                        .join("models--mlx-community--Llama-3.2-1B-Instruct-4bit")
+                        .to_string_lossy()
+                        .into_owned()
+                })
+                .unwrap_or_default();
+
+            resolve_hf_cache_dir(&cache_dir)
+        };
+    }
 
     #[test]
     #[ignore = "requires local model files"]
-    fn test_load_qwen3_model() {
-        let _model = super::load_qwen3_model(CACHED_TEST_MODEL_DIR).unwrap();
+    fn test_load_llama_model() {
+        use mlx_rs::module::ModuleParameters;
+
+        let model_dir = CACHED_TEST_MODEL_DIR.as_str();
+        let cfg = ModelConfig::from_dir(model_dir).unwrap();
+        let Family::Llama(args) = cfg.family else {
+            panic!("expected llama config");
+        };
+        let model = super::Model::new(args).unwrap();
+
+        // Print some model parameter keys
+        let params = model.parameters().flatten();
+        let mut param_keys: Vec<_> = params.keys().map(|k| k.to_string()).collect();
+        param_keys.sort();
+        println!("=== Model parameter keys (first 20) ===");
+        for key in param_keys.iter().take(20) {
+            println!("  {key}");
+        }
+
+        // Print some safetensor keys
+        let weights_path = std::path::Path::new(model_dir).join("model.safetensors");
+        let loaded = mlx_rs::Array::load_safetensors(&weights_path).unwrap();
+        let mut weight_keys: Vec<_> = loaded.keys().map(|k| k.to_string()).collect();
+        weight_keys.sort();
+        println!("=== Safetensor weight keys (first 20) ===");
+        for key in weight_keys.iter().take(20) {
+            println!("  {key}");
+        }
+
+        // Find unmatched keys
+        let param_set: std::collections::HashSet<_> = param_keys.iter().collect();
+        let weight_set: std::collections::HashSet<_> = weight_keys.iter().collect();
+        let unloaded: Vec<_> = weight_set.difference(&param_set).collect();
+        let missing: Vec<_> = param_set.difference(&weight_set).collect();
+        println!(
+            "=== Weight keys NOT in model params ({}) ===",
+            unloaded.len()
+        );
+        for key in unloaded.iter().take(10) {
+            println!("  {key}");
+        }
+        println!(
+            "=== Model param keys NOT in weights ({}) ===",
+            missing.len()
+        );
+        for key in missing.iter().take(10) {
+            println!("  {key}");
+        }
+        println!(
+            "Total model params: {}, Total weight keys: {}",
+            param_keys.len(),
+            weight_keys.len()
+        );
     }
 
-    const CACHED_QUANT_TEST_MODEL_DIR: &str = "../cache/Qwen3-1.7B-MLX-8bit";
+    #[test]
+    #[ignore = "requires local model files"]
+    fn test_load_tokenizer() {
+        let tokenizer = load_llama_tokenizer(CACHED_TEST_MODEL_DIR.as_str()).unwrap();
+
+        let _encoding = tokenizer.encode("Hello, world!", true).unwrap();
+    }
 
     /// Regression guard for the quantised-checkpoint loader path: without
     /// `try_into_quantized` before `load_safetensors`, the packed-uint32
@@ -513,8 +589,10 @@ mod tests {
     /// forward fires `[rms_norm] weight has K elements but x's last dim is D`.
     #[test]
     #[ignore = "requires local quantised model files"]
-    fn quantized_qwen3_model_loads_and_forwards() {
-        let mut model = super::load_qwen3_model(CACHED_QUANT_TEST_MODEL_DIR).unwrap();
+    fn quantized_llama_model_loads_and_forwards() {
+        use mlx_rs::module::Module;
+
+        let mut model = load_llama_model(CACHED_QUANT_TEST_MODEL_DIR.as_str()).unwrap();
         let prompt = Array::from_slice(&[1i32, 2, 3, 4], &[1, 4]);
         let mut cache: Vec<Option<KVCache>> = Vec::new();
         let input = super::ModelInput {
@@ -529,32 +607,25 @@ mod tests {
 
     #[test]
     #[ignore = "requires local model files"]
-    fn test_load_tokenizer() {
-        let tokenizer = load_qwen3_tokenizer(CACHED_TEST_MODEL_DIR).unwrap();
-
-        let _encoding = tokenizer.encode("Hello, world!", true).unwrap();
-    }
-
-    #[test]
-    #[ignore = "requires local model files"]
-    fn test_load_and_run_qwen3_via_generate() {
+    fn test_load_and_run_llama_via_generate() {
         use std::ops::ControlFlow;
 
         use crate::model_context::{generate, load, GenerateParams};
         use crate::sampler::Sampler;
         use crate::user_input::UserInput;
 
-        let mut ctx = load(CACHED_TEST_MODEL_DIR).unwrap();
+        let mut ctx = load(CACHED_TEST_MODEL_DIR.as_str()).unwrap();
+        let input = UserInput::text("What is the capital of France?");
         let params = GenerateParams {
-            max_new_tokens: 10,
+            max_new_tokens: 50,
             sampling: Sampler::Greedy,
             ..Default::default()
         };
-        let result = generate(&mut ctx, UserInput::text("hello"), params, &mut |_, _| {
+        let result = generate(&mut ctx, input, params, &mut |_, _| {
             ControlFlow::Continue(())
         })
         .unwrap();
-        println!("{}", result.text);
+        println!("Response: {}", result.text);
         assert!(result.completion_tokens > 0);
     }
 }
