--- mlx-lm/src/models/mod.rs
@@ -1,2 +1,0 @@
-pub mod llama;
-pub mod qwen3;
