+++ mlx-lm/src/qwen3/mod.rs
@@ -1,6 +1,17 @@
-//! Qwen3 family: the [`crate::LanguageModel`] adapter over the qwen3
-//! model graph in [`crate::models::qwen3`].
+//! Qwen3 text family.
 
-mod adapter;
+pub mod text;
 
-pub use adapter::load_context;
+use std::path::Path;
+
+use crate::config::ModelConfig;
+use crate::error::Error;
+use crate::family::LoadedContext;
+
+pub(crate) fn load_context(cfg: &ModelConfig, dir: &Path) -> Result<LoadedContext, Error> {
+    let env = cfg
+        .family
+        .as_qwen3()
+        .ok_or_else(|| Error::config("qwen3::load_context: wrong family"))?;
+    text::adapter::load_context(cfg, env, dir)
+}
