+++ mlx-lm/src/qwen3/text/mod.rs
@@ -1,0 +1,5 @@
+//! Qwen3 text path: config, model graph, and the runtime adapter.
+
+pub mod adapter;
+pub mod config;
+pub mod model;
