+++ mlx-lm/src/qwen3/text/model.rs
@@ -20,44 +20,15 @@ use crate::{
     cache::KeyValueCache,
     config::{Family, ModelConfig},
     error::Error,
-    family::EosSpec,
     loader::apply_post_load_memory_policy,
     nn::{ensure_cache_populated, AttentionInput, ModelInput, SwigluMlp},
+    qwen3::text::config::ModelArgs,
     utils::{
-        rope::{initialize_rope, FloatOrString, RopeVariant},
-        scaled_dot_product_attention,
+        create_attention_mask,
+        rope::{initialize_rope, RopeVariant},
+        scaled_dot_product_attention, AttentionMask,
     },
 };
-
-/// Llama config body. `model_type` is the serde tag on
-/// [`crate::config::Family`]; quantization lives on the outer
-/// [`crate::config::ModelConfig`].
-#[derive(Debug, Clone, Deserialize)]
-pub struct ModelArgs {
-    pub hidden_size: i32,
-    pub num_hidden_layers: i32,
-    pub intermediate_size: i32,
-    pub num_attention_heads: i32,
-    pub rms_norm_eps: f32,
-    pub vocab_size: i32,
-    pub num_key_value_heads: i32,
-    pub max_position_embeddings: i32,
-    pub rope_theta: f32,
-    pub head_dim: i32,
-    #[serde(default = "default_true")]
-    pub tie_word_embeddings: bool,
-    #[serde(default)]
-    pub attention_bias: bool,
-    #[serde(default)]
-    pub mlp_bias: bool,
-    pub rope_scaling: Option<HashMap<String, FloatOrString>>,
-    #[serde(default)]
-    pub eos_token_id: Option<EosSpec>,
-}
-
-fn default_true() -> bool {
-    true
-}
 
 #[derive(Debug, Clone, ModuleParameters, Quantizable)]
 pub struct Attention {
@@ -78,6 +49,10 @@ pub struct Attention {
     #[param]
     pub o_proj: MaybeQuantized<nn::Linear>,
     #[param]
+    pub q_norm: nn::RmsNorm,
+    #[param]
+    pub k_norm: nn::RmsNorm,
+    #[param]
     pub rope: RopeVariant,
 }
 
@@ -91,16 +66,23 @@ impl Attention {
         let scale = (head_dim as f32).sqrt().recip();
 
         let q_proj = nn::LinearBuilder::new(dim, n_heads * head_dim)
-            .bias(args.attention_bias)
+            .bias(false)
             .build()?;
         let k_proj = nn::LinearBuilder::new(dim, n_kv_heads * head_dim)
-            .bias(args.attention_bias)
+            .bias(false)
             .build()?;
         let v_proj = nn::LinearBuilder::new(dim, n_kv_heads * head_dim)
-            .bias(args.attention_bias)
+            .bias(false)
             .build()?;
         let o_proj = nn::LinearBuilder::new(n_heads * head_dim, dim)
-            .bias(args.attention_bias)
+            .bias(false)
+            .build()?;
+
+        let q_norm = nn::RmsNormBuilder::new(head_dim)
+            .eps(args.rms_norm_eps)
+            .build()?;
+        let k_norm = nn::RmsNormBuilder::new(head_dim)
+            .eps(args.rms_norm_eps)
             .build()?;
 
         let rope = initialize_rope(
@@ -119,6 +101,8 @@ impl Attention {
             k_proj: MaybeQuantized::Original(k_proj),
             v_proj: MaybeQuantized::Original(v_proj),
             o_proj: MaybeQuantized::Original(o_proj),
+            q_norm,
+            k_norm,
             rope,
         })
     }
@@ -126,7 +110,7 @@ impl Attention {
 
 impl<C> Module<AttentionInput<'_, C>> for Attention
 where
-    C: KeyValueCache,
+    C: KeyValueCache + Default,
 {
     type Output = Array;
 
@@ -144,12 +128,16 @@ where
         let keys = self.k_proj.forward(x)?;
         let values = self.v_proj.forward(x)?;
 
-        let mut queries = queries
-            .reshape(&[B, L, self.n_heads, -1])?
-            .transpose_axes(&[0, 2, 1, 3])?;
-        let mut keys = keys
-            .reshape(&[B, L, self.n_kv_heads, -1])?
-            .transpose_axes(&[0, 2, 1, 3])?;
+        let mut queries = self.q_norm.forward(
+            &queries
+                .reshape(&[B, L, self.n_heads, -1])?
+                .transpose_axes(&[0, 2, 1, 3])?,
+        )?;
+        let mut keys = self.k_norm.forward(
+            &keys
+                .reshape(&[B, L, self.n_kv_heads, -1])?
+                .transpose_axes(&[0, 2, 1, 3])?,
+        )?;
         let values = values
             .reshape(&[B, L, self.n_kv_heads, -1])?
             .transpose_axes(&[0, 2, 1, 3])?;
@@ -175,6 +163,8 @@ where
         self.k_proj.training_mode(mode);
         self.v_proj.training_mode(mode);
         self.o_proj.training_mode(mode);
+        self.q_norm.training_mode(mode);
+        self.k_norm.training_mode(mode);
         <RopeVariant as Module<nn::RopeInput>>::training_mode(&mut self.rope, mode);
     }
 }
@@ -205,7 +195,7 @@ impl TransformerBlock {
         let hidden_size = args.hidden_size;
 
         let self_attn = Attention::new(args)?;
-        let mlp = SwigluMlp::new(args.hidden_size, args.intermediate_size, args.mlp_bias)?;
+        let mlp = SwigluMlp::new(args.hidden_size, args.intermediate_size, false)?;
         let input_layernorm = nn::RmsNormBuilder::new(args.hidden_size)
             .eps(args.rms_norm_eps)
             .build()?;
@@ -226,7 +216,7 @@ impl TransformerBlock {
 
 impl<C> Module<AttentionInput<'_, C>> for TransformerBlock
 where
-    C: KeyValueCache,
+    C: KeyValueCache + Default,
 {
     type Output = Array;
 
@@ -258,7 +248,7 @@ where
 }
 
 #[derive(Debug, Clone, ModuleParameters, Quantizable)]
-pub struct LlamaModel {
+pub struct Qwen3Model {
     pub vocab_size: i32,
     pub num_hidden_layers: i32,
 
@@ -274,7 +264,7 @@ pub struct LlamaModel {
     pub norm: nn::RmsNorm,
 }
 
-impl LlamaModel {
+impl Qwen3Model {
     pub fn new(args: &ModelArgs) -> Result<Self, Exception> {
         assert!(args.vocab_size.is_positive());
 
@@ -299,7 +289,7 @@ impl LlamaModel {
     }
 }
 
-impl<C> Module<ModelInput<'_, C>> for LlamaModel
+impl<C> Module<ModelInput<'_, C>> for Qwen3Model
 where
     C: KeyValueCache + Default,
 {
@@ -318,15 +308,13 @@ where
 
         let mask = match mask {
             Some(mask) => Some(mask.clone()),
-            None => {
-                if h.shape()[1] > 1 {
-                    let m =
-                        nn::MultiHeadAttention::create_additive_causal_mask::<f32>(h.shape()[1])?;
-                    Some(m.as_dtype(h.dtype())?)
-                } else {
-                    None
+            None => match create_attention_mask(&h, cache, Some(true))? {
+                Some(AttentionMask::Array(a)) => Some(a),
+                Some(AttentionMask::Causal) => {
+                    return Err(Exception::custom("Only `Array` mask is supported"));
                 }
-            }
+                None => None,
+            },
         };
 
         ensure_cache_populated(cache, self.layers.len());
@@ -358,7 +346,7 @@ pub struct Model {
 
     #[quantizable]
     #[param]
-    pub model: LlamaModel,
+    pub model: Qwen3Model,
 
     #[quantizable]
     #[param]
@@ -367,7 +355,7 @@ pub struct Model {
 
 impl Model {
     pub fn new(args: ModelArgs) -> Result<Self, Exception> {
-        let model = LlamaModel::new(&args)?;
+        let model = Qwen3Model::new(&args)?;
         let lm_head = if !args.tie_word_embeddings {
             Some(MaybeQuantized::Original(
                 nn::LinearBuilder::new(args.hidden_size, args.vocab_size)
@@ -407,14 +395,14 @@ where
     }
 
     fn training_mode(&mut self, mode: bool) {
-        <LlamaModel as Module<ModelInput<'_, C>>>::training_mode(&mut self.model, mode);
+        <Qwen3Model as Module<ModelInput<'_, C>>>::training_mode(&mut self.model, mode);
         if let Some(lm_head) = &mut self.lm_head {
             lm_head.training_mode(mode);
         }
     }
 }
 
-pub fn load_llama_tokenizer(model_dir: impl AsRef<Path>) -> Result<Tokenizer, Error> {
+pub fn load_qwen3_tokenizer(model_dir: impl AsRef<Path>) -> Result<Tokenizer, Error> {
     let file = model_dir.as_ref().join("tokenizer.json");
     Tokenizer::from_file(file).map_err(Into::into)
 }
@@ -425,12 +413,12 @@ pub struct WeightMap {
     pub weight_map: HashMap<String, String>,
 }
 
-pub fn load_llama_model(model_dir: impl AsRef<Path>) -> Result<Model, Error> {
+pub fn load_qwen3_model(model_dir: impl AsRef<Path>) -> Result<Model, Error> {
     let model_dir = model_dir.as_ref();
     let cfg = ModelConfig::from_dir(model_dir)?;
     let quant = cfg.quantization().cloned();
-    let Family::Llama(args) = cfg.family else {
-        return Err(Error::config("config.json model_type is not llama"));
+    let Family::Qwen3(args) = cfg.family else {
+        return Err(Error::config("config.json model_type is not qwen3"));
     };
     let mut model = Model::new(args)?;
     if let Some(q) = quant {
@@ -484,134 +472,19 @@ pub fn load_llama_model(model_dir: impl AsRef<Path>) -> Result<Model, Error> {
 
 #[cfg(test)]
 mod tests {
-    use std::{env::home_dir, fs};
+    use mlx_rs::{module::Module, transforms::eval, Array};
 
-    use lazy_static::lazy_static;
-    use mlx_rs::{transforms::eval, Array};
+    use crate::{cache::KVCache, qwen3::text::model::load_qwen3_tokenizer};
 
-    use crate::{
-        cache::KVCache,
-        config::{Family, ModelConfig},
-        models::llama::{load_llama_model, load_llama_tokenizer},
-    };
-
-    /// Resolve the HuggingFace cache directory to the actual snapshot path.
-    /// The structure is:
-    ///   models--<org>--<name>/
-    ///     refs/
-    ///       main  (contains the commit hash)
-    ///     snapshots/
-    ///       <commit_hash>/  (actual model files)
-    fn resolve_hf_cache_dir(model_cache_dir: &str) -> String {
-        let refs_main = std::path::Path::new(model_cache_dir)
-            .join("refs")
-            .join("main");
-        let commit_hash = fs::read_to_string(&refs_main)
-            .unwrap_or_default()
-            .trim()
-            .to_string();
-        std::path::Path::new(model_cache_dir)
-            .join("snapshots")
-            .join(commit_hash)
-            .to_string_lossy()
-            .into_owned()
-    }
-
-    lazy_static! {
-        static ref CACHED_TEST_MODEL_DIR: String = {
-            let cache_dir = home_dir()
-                .map(|p| {
-                    p.join(".cache")
-                        .join("huggingface")
-                        .join("hub")
-                        .join("models--meta-llama--Llama-3.2-1B-Instruct")
-                        .to_string_lossy()
-                        .into_owned()
-                })
-                .unwrap_or_default();
-
-            resolve_hf_cache_dir(&cache_dir)
-        };
-        static ref CACHED_QUANT_TEST_MODEL_DIR: String = {
-            let cache_dir = home_dir()
-                .map(|p| {
-                    p.join(".cache")
-                        .join("huggingface")
-                        .join("hub")
-                        .join("models--mlx-community--Llama-3.2-1B-Instruct-4bit")
-                        .to_string_lossy()
-                        .into_owned()
-                })
-                .unwrap_or_default();
-
-            resolve_hf_cache_dir(&cache_dir)
-        };
-    }
+    const CACHED_TEST_MODEL_DIR: &str = "../cache/Qwen3-4B-bf16";
 
     #[test]
     #[ignore = "requires local model files"]
-    fn test_load_llama_model() {
-        use mlx_rs::module::ModuleParameters;
-
-        let model_dir = CACHED_TEST_MODEL_DIR.as_str();
-        let cfg = ModelConfig::from_dir(model_dir).unwrap();
-        let Family::Llama(args) = cfg.family else {
-            panic!("expected llama config");
-        };
-        let model = super::Model::new(args).unwrap();
-
-        // Print some model parameter keys
-        let params = model.parameters().flatten();
-        let mut param_keys: Vec<_> = params.keys().map(|k| k.to_string()).collect();
-        param_keys.sort();
-        println!("=== Model parameter keys (first 20) ===");
-        for key in param_keys.iter().take(20) {
-            println!("  {key}");
-        }
-
-        // Print some safetensor keys
-        let weights_path = std::path::Path::new(model_dir).join("model.safetensors");
-        let loaded = mlx_rs::Array::load_safetensors(&weights_path).unwrap();
-        let mut weight_keys: Vec<_> = loaded.keys().map(|k| k.to_string()).collect();
-        weight_keys.sort();
-        println!("=== Safetensor weight keys (first 20) ===");
-        for key in weight_keys.iter().take(20) {
-            println!("  {key}");
-        }
-
-        // Find unmatched keys
-        let param_set: std::collections::HashSet<_> = param_keys.iter().collect();
-        let weight_set: std::collections::HashSet<_> = weight_keys.iter().collect();
-        let unloaded: Vec<_> = weight_set.difference(&param_set).collect();
-        let missing: Vec<_> = param_set.difference(&weight_set).collect();
-        println!(
-            "=== Weight keys NOT in model params ({}) ===",
-            unloaded.len()
-        );
-        for key in unloaded.iter().take(10) {
-            println!("  {key}");
-        }
-        println!(
-            "=== Model param keys NOT in weights ({}) ===",
-            missing.len()
-        );
-        for key in missing.iter().take(10) {
-            println!("  {key}");
-        }
-        println!(
-            "Total model params: {}, Total weight keys: {}",
-            param_keys.len(),
-            weight_keys.len()
-        );
+    fn test_load_qwen3_model() {
+        let _model = super::load_qwen3_model(CACHED_TEST_MODEL_DIR).unwrap();
     }
 
-    #[test]
-    #[ignore = "requires local model files"]
-    fn test_load_tokenizer() {
-        let tokenizer = load_llama_tokenizer(CACHED_TEST_MODEL_DIR.as_str()).unwrap();
-
-        let _encoding = tokenizer.encode("Hello, world!", true).unwrap();
-    }
+    const CACHED_QUANT_TEST_MODEL_DIR: &str = "../cache/Qwen3-1.7B-MLX-8bit";
 
     /// Regression guard for the quantised-checkpoint loader path: without
     /// `try_into_quantized` before `load_safetensors`, the packed-uint32
@@ -619,10 +492,8 @@ mod tests {
     /// forward fires `[rms_norm] weight has K elements but x's last dim is D`.
     #[test]
     #[ignore = "requires local quantised model files"]
-    fn quantized_llama_model_loads_and_forwards() {
-        use mlx_rs::module::Module;
-
-        let mut model = load_llama_model(CACHED_QUANT_TEST_MODEL_DIR.as_str()).unwrap();
+    fn quantized_qwen3_model_loads_and_forwards() {
+        let mut model = super::load_qwen3_model(CACHED_QUANT_TEST_MODEL_DIR).unwrap();
         let prompt = Array::from_slice(&[1i32, 2, 3, 4], &[1, 4]);
         let mut cache: Vec<Option<KVCache>> = Vec::new();
         let input = super::ModelInput {
@@ -637,25 +508,32 @@ mod tests {
 
     #[test]
     #[ignore = "requires local model files"]
-    fn test_load_and_run_llama_via_generate() {
+    fn test_load_tokenizer() {
+        let tokenizer = load_qwen3_tokenizer(CACHED_TEST_MODEL_DIR).unwrap();
+
+        let _encoding = tokenizer.encode("Hello, world!", true).unwrap();
+    }
+
+    #[test]
+    #[ignore = "requires local model files"]
+    fn test_load_and_run_qwen3_via_generate() {
         use std::ops::ControlFlow;
 
         use crate::model_context::{generate, load, GenerateParams};
         use crate::sampler::Sampler;
         use crate::user_input::UserInput;
 
-        let mut ctx = load(CACHED_TEST_MODEL_DIR.as_str()).unwrap();
-        let input = UserInput::text("What is the capital of France?");
+        let mut ctx = load(CACHED_TEST_MODEL_DIR).unwrap();
         let params = GenerateParams {
-            max_new_tokens: 50,
+            max_new_tokens: 10,
             sampling: Sampler::Greedy,
             ..Default::default()
         };
-        let result = generate(&mut ctx, input, params, &mut |_, _| {
+        let result = generate(&mut ctx, UserInput::text("hello"), params, &mut |_, _| {
             ControlFlow::Continue(())
         })
         .unwrap();
-        println!("Response: {}", result.text);
+        println!("{}", result.text);
         assert!(result.completion_tokens > 0);
     }
 }
